# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bazar AI API
  version: 0.2.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  description: |
    ## Контракт
    - Тело: один JSON-объект; лишние поля отклоняются (`DisallowUnknownFields`).
    - Ошибки: обёртка `{"error":{"code","message","request_id"?}}` ([ErrorEnvelope](#/components/schemas/ErrorEnvelope)).
    - Заголовки: ответ содержит `X-Request-Id`; при необходимости передавайте свой. Для refresh — `X-Device-Id` (опционально, должен совпадать при ротации).
    - Публичный заказ: заголовок `Idempotency-Key` (опционально) пробрасывается в заказ.
    - Лимит тела JSON: переменная окружения `API_MAX_BODY_BYTES` (по умолчанию 1 MiB). `multipart/*` на этом middleware не режется — отдельные лимиты в обработчиках (например 5 MiB для картинок).
    - Rate limit: **120** запросов / минуту с одного IP (код `rate_limited`, HTTP **429**).
    - CORS: если `ALLOWED_ORIGINS` не `*`, запросы с неразрешённого `Origin` → **403** `forbidden`.
...
servers:
  - url: http://localhost:8080
    description: Локальная разработка

# По умолчанию все операции требуют Bearer JWT; публичные переопределяют через security: []
security:
  - bearerAuth: []
paths:
...
  /api/v1/stores:
    post:
      operationId: createStore
      tags: [Stores]
      summary: Create store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreCreateRequest"
      responses:
        "201":
          description: Магазин создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/products/{id}/images:
    post:
      operationId: uploadProductImage
      tags: [Products]
      summary: Upload product image (multipart, поле `image`)
      description: Лимит тела ~5 MiB; глобальный JSON MaxBytes на этом маршруте не применяется.
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "201":
          description: URL загруженного изображения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Коды из API, например validation_error, unauthorized, forbidden, not_found,
                conflict, quota_exceeded, rate_limited, payload_too_large, upstream_error, internal_error.
            message:
              type: string
            request_id:
              type: string
...
    StoreCreateRequest:
      allOf:
        - $ref: "#/components/schemas/Store"
      description: |
        Декод как [platform.Store](internal/platform). `owner_id` в теле перезаписывается из JWT.
        Сервер валидирует имя и контакты.
...
    Contacts:
      type: object
      properties:
        phone: { type: string }
        whatsapp: { type: string }
        telegram: { type: string }
        instagram: { type: string }
        address: { type: string }
        schedule: { type: string }
    Store:
      type: object
      properties:
        id: { type: string }
        owner_id: { type: string }
        name: { type: string }
        slug: { type: string }
        niche: { type: string }
        description: { type: string }
        region: { type: string }
        city: { type: string }
        currency: { type: string }
        status: { type: string }
        tariff: { type: string }
        theme: { type: string }
        style: { type: string }
        contacts:
          $ref: "#/components/schemas/Contacts"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ImageUploadResult:
      type: object
      required: [url, preview_url, mime, filename, max_size]
      properties:
        url: { type: string }
        preview_url: { type: string }
        mime: { type: string }
        filename: { type: string }
        max_size:
          type: integer
          description: Максимальный размер в байтах (5 MiB)
...
  responses:
    ValidationError:
      description: Невалидное тело или бизнес-валидация
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: validation_error
              message: invalid payload
    Unauthorized:
      description: Нет или неверный Bearer / сессия
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: missing token
...
    TooManyRequests:
      description: Rate limit 120/мин на IP
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: rate_limited
              message: rate limit exceeded
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: internal_error
              message: unexpected server error
    PayloadTooLarge:
      description: Тело запроса превышает лимит (см. API_MAX_BODY_BYTES); для multipart действуют лимиты обработчика.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: payload_too_large
              message: request body too large